feat: resolve StorageUtils issue & add tests
