+++ PrebidMobile/Utils/StorageUtils.swift
@@ -1,0 +1,34 @@
+/*   Copyright 2018-2019 Prebid.org, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import Foundation
+
+class StorageUtils {
+    
+    static let PB_SharedIdKey = "kPBSharedId"
+    
+    // SharedId
+    static var sharedId: String? {
+        get { UserDefaults.standard.string(forKey: PB_SharedIdKey) }
+        
+        set {
+            if let newValue {
+                UserDefaults.standard.set(newValue, forKey: PB_SharedIdKey)
+            } else {
+                UserDefaults.standard.removeObject(forKey: PB_SharedIdKey)
+            }
+        }
+    }
+}
